+++ src/AgentFrameworkToolkit.Cohere/ServiceCollectionExtensions.cs
@@ -1,0 +1,33 @@
+using JetBrains.Annotations;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace AgentFrameworkToolkit.Cohere;
+
+/// <summary>
+/// Extension Methods for IServiceCollection
+/// </summary>
+[PublicAPI]
+public static class ServiceCollectionExtensions
+{
+    /// <summary>
+    /// Register a CohereAgentFactory as a Singleton
+    /// </summary>
+    /// <param name="services">The IServiceCollection collection</param>
+    /// <param name="connection">Connection Details</param>
+    /// <returns>The ServiceCollection</returns>
+    public static IServiceCollection AddCohereAgentFactory(this IServiceCollection services, CohereConnection connection)
+    {
+        return services.AddSingleton(new CohereAgentFactory(connection));
+    }
+
+    /// <summary>
+    /// Register a CohereAgentFactory as a Singleton
+    /// </summary>
+    /// <param name="services">The IServiceCollection collection</param>
+    /// <param name="apiKey">The API Key</param>
+    /// <returns>The ServiceCollection</returns>
+    public static IServiceCollection AddCohereAgentFactory(this IServiceCollection services, string apiKey)
+    {
+        return services.AddSingleton(new CohereAgentFactory(apiKey));
+    }
+}